Fix mapping type name
Due to previous mismatch between version of elastic search instance
(6.7.1) and elasticsearch-dsl version (7.0.0) and the fact that
saving a document in non-existent index creates an index inferred from
the document, there's a mismatch between mapping type name. This commit
makes it possible to downgrade elasticsearch-dsl to version compatible
with elastic search 6.7.1 without a need for recreating the indexes.