Make ReferenceStore caching optional
The current caching mechanism is useful when processing thousands of papers at once, as all the references can be merge in-memory before saving. The main usage is for prepopulating the store with references. However, when run in parallel, the caching results in inconsistencies and elastic `ConflictError`s due to versions mismatches.